+++ cmn-data-svc-client/src/main/java/org/acumos/cds/CCDSConstants.java
@@ -25,8 +25,6 @@ package org.acumos.cds;
  */
 public class CCDSConstants {
 
-	public static final String X_REQUEST_ID = "X-Request-ID";
-
 	public static final String JUNCTION_QUERY_PARAM = "_j";
 
 	// Metadata